chore: move client.ts to client/v1/index.ts
