+++ src/colab/consumption/notifier.unit.test.ts
@@ -10,7 +10,7 @@ import { telemetry } from '../../telemetry';
 import { LowBalanceSeverity } from '../../telemetry/api';
 import { TestEventEmitter } from '../../test/helpers/events';
 import { newVsCodeStub, VsCodeStub } from '../../test/helpers/vscode';
-import { SubscriptionTier, ConsumptionUserInfo } from '../api';
+import { SubscriptionTier, ConsumptionUserInfo } from '../client/v1/api';
 import { ConsumptionNotifier } from './notifier';
 
 const NOTIFICATION_SEVERITIES = ['warn', 'error'] as const;
