+++ src/colab/consumption/status-bar.ts
@@ -7,7 +7,7 @@
 import vscode, { Disposable, Event, StatusBarItem } from 'vscode';
 import { log } from '../../common/logging';
 import { Toggleable } from '../../common/toggleable';
-import { ConsumptionUserInfo, SubscriptionTier } from '../api';
+import { ConsumptionUserInfo, SubscriptionTier } from '../client/v1/api';
 
 /**
  * Monitors {@link ConsumptionUserInfo} and maintains a {@link StatusBarItem}.
