+++ src/colab/content-browser/content-tree.vscode.test.ts
@@ -21,7 +21,7 @@ import {
 import { ContentsFileSystemProvider } from '../../jupyter/contents/file-system';
 import { ColabAssignedServer } from '../../jupyter/servers';
 import { TestEventEmitter } from '../../test/helpers/events';
-import { Variant } from '../api';
+import { Variant } from '../client/v1/api';
 import {
   COLAB_CLIENT_AGENT_HEADER,
   COLAB_RUNTIME_PROXY_TOKEN_HEADER,
