+++ src/colab/files.unit.test.ts
@@ -9,7 +9,7 @@ import { expect } from 'chai';
 import { describe } from 'mocha';
 import { TestUri } from '../test/helpers/uri';
 import { newVsCodeStub, VsCodeStub } from '../test/helpers/vscode';
-import { Variant } from './api';
+import { Variant } from './client/v1/api';
 import { buildColabFileUri } from './files';
 
 const DEFAULT_SERVER = {
