+++ src/jupyter/client/index.unit.test.ts
@@ -8,7 +8,7 @@ import { randomUUID } from 'crypto';
 import { expect } from 'chai';
 import * as sinon from 'sinon';
 import { Disposable } from 'vscode';
-import { Variant } from '../../colab/api';
+import { Variant } from '../../colab/client/v1/api';
 import {
   COLAB_CLIENT_AGENT_HEADER,
   COLAB_RUNTIME_PROXY_TOKEN_HEADER,
