+++ src/telemetry/telemetry.unit.test.ts
@@ -11,7 +11,7 @@ import { Disposable } from 'vscode';
 import {
   AuthType,
   SubscriptionTier as ColabSubscriptionTier,
-} from '../colab/api';
+} from '../colab/client/v1/api';
 import { COLAB_EXT_IDENTIFIER } from '../config/constants';
 import { JUPYTER_EXT_IDENTIFIER } from '../jupyter/jupyter-extension';
 import { newVsCodeStub, VsCodeStub } from '../test/helpers/vscode';
